require C++17 and deprecate the rcppmath namespace
